Revert "bugfix for profile screen"
