+++ lib/Screens/ProfileScreen/profile_screen.dart
@@ -12,7 +12,7 @@ class _ProfileScreenState extends State<ProfileScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Body(),
+      body: SafeArea(child: Body()),
     );
   }
 }
